+++ test/index.test.js
@@ -281,7 +281,7 @@ test('Should throw when trying to connect on an invalid host', async (t) => {
   t.after(() => fastify.close())
 
   fastify.register(fastifyValkey, {
-    addresses: [{ host: 'invalid_host' }],
+    addresses: [{ host: 'invalid_host', port: 9999 }],
     connectionBackoff: {
       numberOfRetries: 0
     }
@@ -315,7 +315,7 @@ test('Should be able to register multiple namespaced @fastify/valkey instances',
 test('Should throw when @fastify/valkey is initialized with an option that makes valkey throw', async (t) => {
   t.plan(1)
 
-  const fastify = Fastify()
+  const fastify = Fastify({ pluginTimeout: 20000 })
   t.after(() => fastify.close())
 
   fastify.register(fastifyValkey, { addresses: [] })
@@ -326,7 +326,7 @@ test('Should throw when @fastify/valkey is initialized with an option that makes
 test('Should throw when @fastify/valkey is initialized with a namespace and an option that makes valkey throw', async (t) => {
   t.plan(1)
 
-  const fastify = Fastify()
+  const fastify = Fastify({ pluginTimeout: 20000 })
   t.after(() => fastify.close())
 
   fastify.register(fastifyValkey, {
